requirements.txt
-i https://pypi.org/simple
anyio==3.7.1; python_version >= '3.7'
asgiref==3.7.2; python_version >= '3.7'
bcrypt==3.2.0; python_version >= '3.6'
cffi==1.16.0; python_version >= '3.8'
click==8.1.7; python_version >= '3.7'
dnspython==2.1.0; python_version >= '3.6'
fastapi==0.70.0; python_full_version >= '3.6.1'
fastapi-csrf-protect==0.2.1; python_full_version >= '3.6.1' and python_full_version < '4.0.0'
gunicorn==20.1.0; python_version >= '3.5'
h11==0.14.0; python_version >= '3.7'
idna==3.6; python_version >= '3.5'
itsdangerous==2.1.2; python_version >= '3.7'
motor==3.3.2; python_version >= '3.7'
passlib==1.7.4
pycparser==2.21
pydantic==1.10.14; python_version >= '3.7'
pyjwt==2.3.0; python_version >= '3.6'
pymongo==4.6.1; python_version >= '3.7'
python-decouple==3.3
setuptools==69.1.0; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.0; python_version >= '3.7'
starlette==0.16.0; python_version >= '3.6'
typing-extensions==4.9.0; python_version >= '3.8'
uvicorn==0.14.0
